jscookie.org domain is not owned by @FagnerMartinsBrack or @carhartl. I used this package definition at first in hope that, if in the future the domain is acquired, then there would be no drastic refactor to the codebase.
As from gh-6, I realized that if Maven Central will not allow the
artifactId to have a domain reference that is not owned by the artifact
owners, then there's no reason to do the same with the package names
throughout the project.

Also, if the user sees a package name starting as com.github.jscookie in
some artifact, then he will know exactly where to look for the project.
