+++ src/main/java/com/github/jscookie/javacookie/ConverterException.java
@@ -1,0 +1,8 @@
+package com.github.jscookie.javacookie;
+
+public final class ConverterException extends Exception {
+	private static final long serialVersionUID = 1;
+	public ConverterException( Throwable cause ) {
+		super( cause );
+	}
+}
