+++ src/main/java/com/github/jscookie/javacookie/CookieParseException.java
@@ -1,0 +1,13 @@
+package com.github.jscookie.javacookie;
+
+public final class CookieParseException extends Exception  {
+	private static final long serialVersionUID = 1;
+	@SuppressWarnings( "unused" )
+	private CookieParseException() {}
+	CookieParseException( Throwable cause ) {
+		super( cause );
+	}
+	CookieParseException( String msg ) {
+		super( msg );
+	}
+}
