+++ src/main/java/com/github/jscookie/javacookie/CookieValue.java
@@ -1,0 +1,7 @@
+package com.github.jscookie.javacookie;
+
+/**
+ * Use this interface to mark that a given class can be used as a cookie value and
+ * serialized/deserialized to/from JSON
+ */
+public interface CookieValue {}
