--- src/main/java/org/jscookie/ConverterException.java
@@ -1,8 +1,0 @@
-package org.jscookie;
-
-public final class ConverterException extends Exception {
-	private static final long serialVersionUID = 1;
-	public ConverterException( Throwable cause ) {
-		super( cause );
-	}
-}
