--- src/main/java/org/jscookie/CookieParseException.java
@@ -1,13 +1,0 @@
-package org.jscookie;
-
-public final class CookieParseException extends Exception  {
-	private static final long serialVersionUID = 1;
-	@SuppressWarnings( "unused" )
-	private CookieParseException() {}
-	CookieParseException( Throwable cause ) {
-		super( cause );
-	}
-	CookieParseException( String msg ) {
-		super( msg );
-	}
-}
