--- src/main/java/org/jscookie/CookieValue.java
@@ -1,7 +1,0 @@
-package org.jscookie;
-
-/**
- * Use this interface to mark that a given class can be used as a cookie value and
- * serialized/deserialized to/from JSON
- */
-public interface CookieValue {}
