+++ src/test/java/com/github/jscookie/javacookie/test/integration/qunit/QUnitPageObject.java
@@ -1,0 +1,54 @@
+package com.github.jscookie.javacookie.test.integration.qunit;
+
+import java.io.IOException;
+
+import org.openqa.selenium.JavascriptExecutor;
+import org.openqa.selenium.WebDriver;
+import org.openqa.selenium.support.ui.WebDriverWait;
+
+import com.fasterxml.jackson.databind.ObjectMapper;
+import com.github.jscookie.javacookie.test.integration.test.utils.Debug;
+import com.google.common.base.Function;
+
+public class QUnitPageObject {
+	private WebDriver driver;
+	private static int TIMEOUT_IN_SECONDS = 1800;
+	private ObjectMapper mapper = new ObjectMapper();
+	public QUnitPageObject( WebDriver driver ) {
+		this.driver = driver;
+	}
+	public QUnitResults waitForTests( final Debug debug ) {
+		return new WebDriverWait( driver, TIMEOUT_IN_SECONDS )
+		.until(new Function<WebDriver, QUnitResults>() {
+			@Override
+			public QUnitResults apply( WebDriver input ) {
+				JavascriptExecutor js;
+				if ( driver instanceof JavascriptExecutor ) {
+					js = ( JavascriptExecutor )driver;
+					String result = ( String )js.executeScript(
+						"return window.global_test_results && JSON.stringify(window.global_test_results)"
+					);
+					if ( debug.is( true ) ) {
+						return null;
+					}
+					System.out.println( "Waiting for 'window.global_test_results': " + result );
+					if ( result == null ) {
+						return null;
+					}
+					try {
+						return mapper.readValue( result, QUnitResults.class );
+					} catch ( IOException cause ) {
+						throw new GlobalTestResultException( result, cause );
+					}
+				}
+				return null;
+			}
+		});
+	}
+	private class GlobalTestResultException extends IllegalStateException {
+		private static final long serialVersionUID = 1;
+		private GlobalTestResultException( String result, Throwable cause ) {
+			super( "Failed to read 'window.global_test_results': " + result, cause );
+		}
+	}
+}
