+++ src/test/java/com/github/jscookie/javacookie/test/integration/test/utils/Debug.java
@@ -1,0 +1,12 @@
+package com.github.jscookie.javacookie.test.integration.test.utils;
+
+public enum Debug {
+	TRUE( true ), FALSE( false );
+	private boolean state;
+	private Debug( boolean state ) {
+		this.state = state;
+	}
+	public boolean is( boolean state ) {
+		return this.state == state;
+	}
+}
