--- src/test/java/org/jscookie/test/unit/utils/BaseTest.java
@@ -1,11 +1,0 @@
-package org.jscookie.test.unit.utils;
-
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
-
-import org.mockito.Mock;
-
-public class BaseTest {
-	protected @Mock HttpServletRequest request;
-	protected @Mock HttpServletResponse response;
-}
